services/embedding-server/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pyannote.audio==3.1.1
torch>=2.1.0
torchaudio>=2.1.0
pydantic==2.6.1
httpx==0.26.0
python-multipart==0.0.9
numpy>=1.24.0
